fix lang hidden bug
