+++ dinky-web/src/global.less
@@ -21,7 +21,8 @@ body,
   height: 100%;
   margin: 0;
   padding: 0;
-  overflow: hidden;
+  overflow-x: auto;
+  overflow-y: auto;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial,
   'Noto Sans', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol',
   'Noto Color Emoji';
@@ -41,15 +42,15 @@ body,
   line-height: 144px;
   position: relative;
   font-weight: 600;
-  font-size: 60px;
-  font-family: Avenir, 'Helvetica Neue', Arial, Helvetica, sans-serif;
+  font-size: 4rem;
+  font-family: PMingLiU-ExtB, 'Helvetica Neue', Arial, Helvetica, sans-serif;
 
 }
 
 .logo {
   position: absolute;
-  width: 30vw;
-  left: 20vw;
+  width: 28vw;
+  left: 15vw;
   display: flex;
   flex: auto ;
   flex-flow: column-reverse;
@@ -57,8 +58,8 @@ body,
 
 
 .desc {
-  margin-bottom: 4vh;
-  margin-top: 2vh;
+  margin-bottom: 2rem;
+  margin-top: 2rem;
   color: @text-color-secondary;
   font-size: @font-size-base;
 }
@@ -67,7 +68,8 @@ body,
 .loginformMain  {
   display: flex;
   flex: auto;
-  padding: 0vh 5vh 30vh 2vh;
+  max-height: max-content;
+  padding: 0vh 6vh 30vh 2vh;
   justify-content: center ;
   align-items: center;
   flex-direction: column;
@@ -75,18 +77,15 @@ body,
   background-position: left center;
   background-repeat: no-repeat;
   z-index: 1;
-  left: 0;
-  .loginform {
-    padding: 0vh 5vh 30vh 2vh;
-  }
-
   .form{
     position: relative;
+    height: 100%;
+    box-shadow: 0 0 10px 5px rgba(0, 0, 0, 0.4);
     margin-left: 60vw;
     margin-bottom: -15vh;
-    height: 60%;
-    min-height: 50vh;
-    box-shadow: 0 0 10px 5px rgba(0, 0, 0, 0.4);
+  }
+  .loginform {
+    padding: 2vh 2vw;
   }
 }
 // .login-form style --- end ----
